emit_i64_andi accepts third parameter as int32_t but the argument being passed is casted to int64_t.
Bug: chromium:340583170
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5538459
Auto-Submit: Zequan Wu <zequanwu@google.com>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93936}

+++ src/wasm/baseline/liftoff-compiler.cc
@@ -8224,7 +8224,7 @@ class LiftoffCompiler {
                            static_cast<int32_t>(~kWeakHeapObjectMask));
         } else {
           __ emit_i64_andi(real_rtt.reg(), real_rtt.reg(),
-                           static_cast<int64_t>(~kWeakHeapObjectMask));
+                           static_cast<int32_t>(~kWeakHeapObjectMask));
         }
         // Constant-time subtyping check: load exactly one candidate RTT from
         // the supertypes list.
